+++ app/src/api/index.ts
@@ -24,7 +24,6 @@ import {
   joinSuccess,
   AutomodSelectionStrategy,
   createStackedMessages,
-  TimerStyle,
 } from '@opentalk/common';
 import {
   AutomodEventType,
@@ -115,7 +114,6 @@ import {
   sharedFolder,
 } from './types/incoming';
 import { Role } from './types/incoming/control';
-import { TimerStopKind } from './types/incoming/timer';
 import { Action as OutgoingActionType, automod } from './types/outgoing';
 import * as outgoing from './types/outgoing';
 import { ClearGlobalMessages } from './types/outgoing/chat';
@@ -839,35 +837,14 @@ const handleProtocolMessage = (dispatch: AppDispatch, data: protocol.IncomingPro
  * @param {AppDispatch} dispatch - this is the dispatch function from the redux store.
  * @param {timer.Message} data Message content
  */
-const handleTimerMessage = (dispatch: AppDispatch, data: timer.Message, state: RootState) => {
+const handleTimerMessage = (dispatch: AppDispatch, data: timer.Message) => {
   switch (data.message) {
     case 'started':
-      if (data.style === TimerStyle.CoffeeBreak) {
-        localMediaContext.reconfigure({ audio: false, video: false });
-      }
       dispatch(startedTimer(data));
       break;
-    case 'stopped': {
-      const timerStyle = state.timer.style;
-      if (timerStyle === TimerStyle.Normal) {
-        switch (data.kind) {
-          case TimerStopKind.Expired:
-            notifications.info(i18next.t('timer-notification-ran-out'));
-            break;
-          case TimerStopKind.ByModerator:
-            notifications.info(i18next.t('timer-notification-stopped'));
-            break;
-          case TimerStopKind.CreatorLeft:
-          default:
-            break;
-        }
-      }
-      if (timerStyle === TimerStyle.CoffeeBreak) {
-        notifications.info(i18next.t('coffee-break-notification'));
-      }
-      dispatch(stoppedTimer());
+    case 'stopped':
+      dispatch(stoppedTimer(data));
       break;
-    }
     case 'updated_ready_status':
       dispatch(updateParticipantsReady(data));
       break;
@@ -1019,7 +996,7 @@ const onMessage =
         handleChatMessage(dispatch, message.payload, message.timestamp, getState());
         break;
       case 'timer':
-        handleTimerMessage(dispatch, message.payload, getState());
+        handleTimerMessage(dispatch, message.payload);
         break;
       case 'whiteboard':
         handleWhiteboardMessage(dispatch, message.payload);