+++ app/src/components/TimerTab/fragments/TimerNormalCounter.tsx
@@ -4,8 +4,8 @@
 import { Box, styled, Typography } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 
-import { useAppSelector } from '../../../hooks';
-import { selectTimerEndsAt, selectRemainingTime } from '../../../store/slices/timerSlice';
+import { useAppSelector, useRemainingDurationOfTimer } from '../../../hooks';
+import { selectTimerEndsAt } from '../../../store/slices/timerSlice';
 import { TimerCounterProps } from './TimerTypes';
 
 const Content = styled(Box)<TimerCounterProps>(({ alignItems = 'initial' }) => ({
@@ -17,14 +17,13 @@ const Content = styled(Box)<TimerCounterProps>(({ alignItems = 'initial' }) => (
 
 const TimerNormalCounter = ({ alignItems }: TimerCounterProps) => {
   const { t } = useTranslation();
-  const realTime = useAppSelector(selectRemainingTime);
+  const remainingTime = useRemainingDurationOfTimer();
   const endTime = useAppSelector(selectTimerEndsAt);
-  const endDate = Boolean(endTime);
 
   return (
     <Content alignItems={alignItems}>
-      <Typography variant="caption">{t(`timer-counter-${endDate ? 'remaining' : 'elapsed'}-time`)}</Typography>
-      <Typography>{realTime?.durationString}</Typography>
+      <Typography variant="caption">{t(`timer-counter-${endTime ? 'remaining' : 'elapsed'}-time`)}</Typography>
+      <Typography>{remainingTime?.durationString}</Typography>
     </Content>
   );
 };