+++ app/src/store/slices/timerSlice.ts
@@ -1,17 +1,14 @@
 // SPDX-FileCopyrightText: OpenTalk GmbH <mail@opentalk.eu>
 //
 // SPDX-License-Identifier: EUPL-1.2
-import { ParticipantId, Timestamp, TimerKind, TimerStyle, joinSuccess } from '@opentalk/common';
-import { createSlice, PayloadAction } from '@reduxjs/toolkit';
+import { ParticipantId, Timestamp, TimerKind, TimerStyle, joinSuccess, notifications } from '@opentalk/common';
+import { createListenerMiddleware, createSlice, PayloadAction, TypedStartListening } from '@reduxjs/toolkit';
 import { intervalToDuration } from 'date-fns';
+import i18next from 'i18next';
 
-import { RootState } from '../';
-import { TimerStarted, ReadyToContinue } from '../../api/types/incoming/timer';
-
-interface RemainingTime {
-  duration: Duration;
-  durationString: string;
-}
+import { AppDispatch, RootState } from '../';
+import { TimerStarted, ReadyToContinue, TimerStopKind, TimerStopped } from '../../api/types/incoming/timer';
+import localMediaContext from '../../modules/Media/LocalMedia';
 
 interface State {
   startedAt?: Timestamp;
@@ -23,8 +20,8 @@ interface State {
   running: boolean;
   kind?: TimerKind;
   style?: TimerStyle;
-  remainingTime?: RemainingTime;
   totalDuration?: Duration;
+  stopTimerKind?: TimerStopKind;
 }
 
 const initialState: State = {
@@ -37,8 +34,8 @@ const initialState: State = {
   running: false,
   kind: undefined,
   style: undefined,
-  remainingTime: undefined,
   totalDuration: undefined,
+  stopTimerKind: undefined,
 };
 
 export const timerSlice = createSlice({
@@ -47,36 +44,34 @@ export const timerSlice = createSlice({
   reducers: {
     resetTimerState: () => initialState,
     startedTimer: (state, { payload }: PayloadAction<TimerStarted>) => {
-      // state = {
-      //   ...payload,
-      //   running: true,
-      //   participantsReady: []
-      // };
+      state.timerId = payload.timerId;
       state.startedAt = payload.startedAt;
       state.endsAt = payload.endsAt;
-      state.timerId = payload.timerId;
       state.readyCheckEnabled = payload.readyCheckEnabled;
       state.title = payload.title;
       state.running = true;
       state.kind = payload.kind;
       state.style = payload.style;
-      state.totalDuration = intervalToDuration({
-        start: new Date(payload.startedAt),
-        end: new Date(payload.endsAt),
-      });
+      if (payload.endsAt) {
+        state.totalDuration = intervalToDuration({
+          start: new Date(payload.startedAt),
+          end: new Date(payload.endsAt),
+        });
+      }
+      state.stopTimerKind = undefined;
     },
-    stoppedTimer: (state) => {
+    stoppedTimer: (state, { payload }: PayloadAction<TimerStopped>) => {
+      state.timerId = undefined;
       state.startedAt = undefined;
       state.endsAt = undefined;
-      state.timerId = undefined;
       state.readyCheckEnabled = undefined;
       state.title = undefined;
       state.running = false;
       state.kind = undefined;
       state.style = undefined;
       state.totalDuration = undefined;
-      state.remainingTime = undefined;
-      // participantsReady: [],
+      state.participantsReady = [];
+      state.stopTimerKind = payload.kind;
     },
     updateParticipantsReady: (state, { payload }: PayloadAction<ReadyToContinue>) => {
       if (payload.status === true && !state.participantsReady.includes(payload.participantId)) {
@@ -87,32 +82,30 @@ export const timerSlice = createSlice({
         state.participantsReady = state.participantsReady.filter((item) => item !== payload.participantId);
       }
     },
-    updateRemainingTime: (state, { payload }: PayloadAction<RemainingTime>) => {
-      state.remainingTime = payload;
-    },
   },
   extraReducers: (builder) => {
     builder.addCase(joinSuccess, (state, { payload: { timer } }) => {
       if (timer) {
-        // state = {
-        //   ...timer,
-        //   running: true,
-        //   participantsReady: []
-        // };
+        state.timerId = timer.timerId;
+        state.startedAt = timer.startedAt;
         state.endsAt = timer.endsAt;
+        state.running = true;
         state.kind = timer.kind;
         state.readyCheckEnabled = timer.readyCheckEnabled;
-        state.startedAt = timer.startedAt;
         state.style = timer.style;
-        state.timerId = timer.timerId;
-        state.running = true;
+        if (timer.endsAt) {
+          state.totalDuration = intervalToDuration({
+            start: new Date(timer.startedAt),
+            end: new Date(timer.endsAt),
+          });
+        }
+        state.stopTimerKind = undefined;
       }
     });
   },
 });
 
-export const { startedTimer, stoppedTimer, updateParticipantsReady, updateRemainingTime, resetTimerState } =
-  timerSlice.actions;
+export const { startedTimer, stoppedTimer, updateParticipantsReady, resetTimerState } = timerSlice.actions;
 
 export const actions = timerSlice.actions;
 
@@ -126,8 +119,48 @@ export const selectTimerRunning = (state: RootState) => state.timer.running;
 export const selectTimerKind = (state: RootState) => state.timer.kind;
 export const selectTimerStyle = (state: RootState) => state.timer.style;
 export const selectTotalDuration = (state: RootState) => state.timer.totalDuration;
-export const selectRemainingTime = (state: RootState) => state.timer.remainingTime;
 export const selectCoffeeBreakTimerId = (state: RootState) =>
   state.timer.style === TimerStyle.CoffeeBreak ? state.timer.timerId : undefined;
 
 export default timerSlice.reducer;
+
+export const timerMiddleware = createListenerMiddleware();
+type AppStartListening = TypedStartListening<RootState, AppDispatch>;
+
+const startAppListening = timerMiddleware.startListening as AppStartListening;
+
+startAppListening({
+  actionCreator: startedTimer,
+  effect: (action, listenerApi) => {
+    const mediaState = listenerApi.getOriginalState().media;
+
+    //Avoid sending reconfigure if both video and audio tracks are not active
+    const isAnyMediaTrackEnabled = mediaState.audioEnabled || mediaState.videoEnabled;
+    if (action.payload.style === TimerStyle.CoffeeBreak && isAnyMediaTrackEnabled) {
+      localMediaContext.reconfigure({ audio: false, video: false });
+    }
+  },
+});
+
+startAppListening({
+  actionCreator: stoppedTimer,
+  effect: (action, listenerApi) => {
+    const timerStyle = listenerApi.getOriginalState().timer.style;
+    if (timerStyle === TimerStyle.Normal) {
+      switch (action.payload.kind) {
+        case TimerStopKind.Expired:
+          notifications.info(i18next.t('timer-notification-ran-out'));
+          break;
+        case TimerStopKind.ByModerator:
+          notifications.info(i18next.t('timer-notification-stopped'));
+          break;
+        case TimerStopKind.CreatorLeft:
+        default:
+          break;
+      }
+    }
+    if (timerStyle === TimerStyle.CoffeeBreak) {
+      notifications.info(i18next.t('coffee-break-notification'));
+    }
+  },
+});