+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Upgrade test packages to remove unmaintained packaged (!1006)
 - Improve sidebar tab definition, types and accessibility (!997)
 - Fix WebRTC stats error in Safari for missing remote report (!1014)
+- Rework most timer/coffee break related components and slices (!1018)
 
 ### New Features